requirements.txt

wheel

easygui

loguru

python-docx

pandas
numpy
openpyxl

comtypes


tqdm
